New stream renderer!
Uses MIME type jupyter/stream (for now)
